Add encoding functions for waveforms
The functions needed to encode waveforms into the necessary data strings are now
written, but the encodeWaveformElement function is not currently tail-recursive.
Since the WaveformElement type is recursive, this could result in a stack
overflow if the types are nested deeply enough.

This shouldn't be a problem in practice, because there's a limit to how many
bytes can be written in one go to the machine, but it would be nice to have it
in full generality.