Simplify waveform model and encoding
Un-spaghetti-d the encoding code by removing the incorrect recursion from the
model and forcing the function to be much more type-safe.

The previous recursive model was actually linear, allowing repeats at multiple
different levels, though this would only end up with multiplying all the repeats
together and applying them to the single sequence at the bottom of the stack.