Initialise IQData.Decode module
Refactor the IQData module to have submodules IQData.Encode and IQData.Decode so
we will also be able to recieve waveforms and sequences from the machine.
Changed encoding functions names to the convention to[Type], so the semantics
are now IQData.Encode.to[Type].  The Decode module should have a nice analogue.