Add Translate module and whitespace changes
Make the Encode and Decode modules submodules of the larger Translate module,
since the former two will in general be opened together, and are directly
related.

Also changed the public Waveform type to take a string as its name rather than
an ASCII string - the user needn't know the details of how its sent when we can
do the encoding ourselves in the process.