Move 3 segment storing functions into 1 function
Previously the three parts of a segment were stored by separate functions, which
was inefficient due to the need to encode the segment 3 times, and also required
the user to have too much knowledge of the system.  The new way just requires
the syntax ``storeSegmentFiles keysight segment``.

Also fixed a type in the file name for header files, which hadn't been written
before.