Add deletion of files from BBG memory
Added function to allow deletion of single files from the BBG memory of the
device.

Changed `setBytesValue` to take a key as a string rather than as a byte array,
so we can keep it looking the same as every other part of the program, and if
there are ever any keys which can write both byte data and string data, then we
can use them together.