Deprecate the CRC module
The CRC module was deprecated in favour of using F#'s built in hash function,
though the code remains.  The built-in hash is only 32-bit, while the CRC module
did a CRC-64, giving less chance of a hash collision.  If this proves to be a
problem, then we can reimplement the CRC module.

Potentially we could just use the CRC-64 module as a checker if a hash collision
occurs - the module is not as fast as the bulitin hash, but chances of a hash
collision in both types are astronomical.