First draft of dictionary compression
The program now compresses onto the machine, but it takes a considerable time to
do so.  It can theoretically cope with any number of repetitions, however, so
this closes #4.

The compression is not yet adequate for it to be considered as a solution to #2,
but it does not exceed memory on realistic sized experiments, so it does close #3.