Resolve #2 - speed up compression massively
Change a few lists to arrays for speed, and use knowledge of data lengths to
create the array once, and use mutability of the elements to make it fast.  This
reduces a few functions from O(n^2) down to O(n), and massively reduces some
constant factors on others.

Reduced dependence on BitConverter and Array.rev, opting instead for bit-masks
and direct writing of data into arrays.