Hide ID details from the user, make overall Control module
Rearranged several types to hide the idea of a segment or sequence Id from the
user.  Moved the hashing functions to a new module so they're accessible
everywhere, and can thus be used to generate names by everyone.

All control functions in RfPulse and Waveform were moved into a generic Control
module, so they're easily accessible.  The functions storeX were changed to not
take an id, but to generate one and then pass this into the new storeXById
functions.