Add experiment printing functions in DEBUG mode
When the compiler is in Debug mode, some additional functions for printing out
experiments are compiled in, for checking that everything is working correctly
after the compression step.  These can print out the whole experiment as a
human-readable sequence of samples and their lengths, along with the names of
Segments and Sequences that contain them, and in the order they're going to be
written.