Add compilation into separate sequences per point
Compilation now splits up the experiment into separate lists for each point of
an experiment (each phase or repetition). Now writes all of these into lots of
different sequences on the machine.  This is currently very slow, though, since
writing sequences seems to take a non-negligible fraction of a second.

This adds the basis for #6, #19, #20 and #23.