Fix #27 split received errors only on first comma
parseError returns an error if the number of "parts" of the error message was
not 2, and split error messages into parts by splitting on commas. Some SCPI
commands include commas, such as those to create sequences, so when the system
failed on these, there was an unhandled exception.

Now we split only on the first comma, so the check - whether or not there is an
integer error code - is still valid, and the number of parts now accurately
reflects the pairing of [| code; message |].