Add unified type Waveform
Segment and Sequence can now be unified into the type of Waveform.  The
storeSegment and storeSequence families of functions are still accessible by the
user, though this may change in the future.  The two types can be safely unified
in most cases, because as far as the user should see, they should just choose a
file to playback, regardless of how it's stored.